@@ -7,7 +7,10 @@
   "scripts": {
     "build": "tsc -d",
     "build:dev": "tsc -d -w",
-    "lint": "tslint --fix 'src/**/*.ts'",
+    "eslint": "eslint \"**/*.ts\"",
+    "eslint:fix": "eslint --fix \"**/*.ts\"",
+    "prettier": "prettier --config .prettierrc.json --check **/*.ts",
+    "prettier:fix": "prettier --config .prettierrc.json --write **/*.ts",
     "test": "jest --coverage",
     "test:dev": "jest --watch"
   },
@@ -19,13 +22,22 @@
   "license": "ISC",
   "dependencies": {
     "roll-parser": "^2.3.2",
-    "typescript": "^3.8.3"
+    "typescript": "^4.7.4"
   },
   "devDependencies": {
-    "@types/jest": "^25.2.1",
+    "@types/jest": "^28.1.4",
     "@types/node": "^13.13.4",
-    "jest": "^27.5.1",
-    "ts-jest": "^25.4.0",
-    "tslint": "^6.1.2"
+    "@typescript-eslint/eslint-plugin": "^5.30.4",
+    "@typescript-eslint/parser": "^5.30.4",
+    "eslint": "^8.19.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-import-resolver-typescript": "^3.2.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsx-a11y": "^6.6.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-simple-import-sort": "^7.0.0",
+    "jest": "^28.1.2",
+    "ts-jest": "^28.0.5"
   }
 }
