@@ -1,11 +1,5 @@
-"use strict";
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-var Manager_1 = __importDefault(require("./Manager"));
-var vars_1 = require("./vars");
-exports.VARS = vars_1.default;
-var constants_1 = require("./constants");
-exports.CONST = constants_1.default;
-exports.default = Manager_1.default;
+import Manager from './Manager';
+export { default as VARS } from './vars';
+export { default as CONST } from './constants';
+export default Manager;
+//# sourceMappingURL=index.js.map
