@@ -1,13 +1,7 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-var manager_1 = require("../manager");
-var word_1 = require("../word");
-exports.checkFullToken = function (token) {
-    return (token[0] === manager_1.VARS.COMMAND
-        && token[1] === manager_1.VARS.BRACE[0]
-        && token.slice(-1) === manager_1.VARS.BRACE[1]);
-};
-exports.checkSubToken = function (token) {
-    return (token[0] === manager_1.VARS.COMMAND
-        && word_1.SUBTOKENS.indexOf(token.slice(1)) > -1);
-};
+import { VARS } from '../manager';
+import { SUBTOKENS } from '../word';
+export const checkFullToken = (token) => token[0] === VARS.COMMAND &&
+    token[1] === VARS.BRACE[0] &&
+    token.slice(-1) === VARS.BRACE[1];
+export const checkSubToken = (token) => token[0] === VARS.COMMAND && SUBTOKENS.indexOf(token.slice(1)) > -1;
+//# sourceMappingURL=tokeniser.js.map
