@@ -1,49 +1,44 @@
-"use strict";
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-var manager_1 = __importDefault(require("../manager"));
-var utils_1 = require("../utils");
-var errors_1 = require("../errors");
-var Word = /** @class */ (function () {
-    function Word(words, weight) {
-        var _this = this;
-        if (weight === void 0) { weight = 1; }
-        this.generate = function () {
+import { MissingBucketError } from '../errors';
+import Manager from '../manager';
+import { checkFullToken, checkSubToken, splitString } from '../utils';
+class Word {
+    constructor(words, weight = 1) {
+        this.generate = () => {
             var _a;
-            var tokens = utils_1.splitString(_this.words);
-            var _loop_1 = function (token) {
-                if (utils_1.checkFullToken(tokens[token])) {
-                    var fragments = [];
-                    var subTokens = tokens[token].slice(2, -1).split(/, ?/);
-                    var aOrAn = "";
-                    for (var _i = 0, subTokens_1 = subTokens; _i < subTokens_1.length; _i++) {
-                        var subToken = subTokens_1[_i];
-                        if (utils_1.checkSubToken(subToken)) {
+            const tokens = splitString(this.words);
+            // eslint-disable-next-line no-restricted-syntax
+            for (const token in tokens) {
+                if (checkFullToken(tokens[token])) {
+                    const fragments = [];
+                    const subTokens = tokens[token].slice(2, -1).split(/, ?/);
+                    let aOrAn = '';
+                    // eslint-disable-next-line no-restricted-syntax
+                    for (const subToken of subTokens) {
+                        if (checkSubToken(subToken)) {
                             // set flags for special cases here;
                             switch (subToken.slice(1)) {
-                                case "a":
-                                case "an":
+                                case 'a':
+                                case 'an':
                                     aOrAn = subToken.slice(1);
+                                    break;
+                                default:
                                     break;
                             }
                         }
                         else {
                             try {
-                                var word = manager_1.default.generate(subToken);
+                                const word = Manager.generate(subToken);
                                 if (word.length > 0) {
                                     fragments.push(word);
                                 }
                             }
                             catch (e) {
-                                // Jest gets this type wrong, so it isn't tested.
-                                if (typeof e === typeof errors_1.MissingBucketError) {
-                                    // tslint:disable-next-line: no-console
-                                    console.error("Swallowing error:");
-                                    // tslint:disable-next-line: no-console
+                                if (typeof e === typeof MissingBucketError) {
+                                    // eslint-disable-next-line no-console
+                                    console.error('Swallowing error:');
+                                    // eslint-disable-next-line no-console
                                     console.error(e);
-                                    fragments.push("!!! " + e.message + " in " + e.bucket + " !!!");
+                                    fragments.push(`!!! ${e.message} !!!`);
                                 }
                                 else {
                                     throw e;
@@ -51,48 +46,45 @@ var Word = /** @class */ (function () {
                             }
                         }
                     }
-                    var outputPrepend = function (fragment) {
-                        if (output_1.length > 0) {
-                            output_1 = fragment + " " + output_1;
+                    let output = fragments.join(', ');
+                    const outputPrepend = (fragment) => {
+                        if (output.length > 0) {
+                            output = `${fragment} ${output}`;
                         }
                         else {
-                            output_1 = fragment;
+                            output = fragment;
                         }
                     };
-                    var output_1 = fragments.join(", ");
                     if (aOrAn.length > 0) {
-                        var vowelArray = ["a", "e", "i", "o", "u", "1", "8"];
-                        var firstChar = (_a = output_1.match(/[a-zA-Z0-9]/)) === null || _a === void 0 ? void 0 : _a.pop();
+                        const vowelArray = ['a', 'e', 'i', 'o', 'u', '1', '8'];
+                        const firstChar = (_a = output.match(/[a-zA-Z0-9]/)) === null || _a === void 0 ? void 0 : _a.pop();
                         if (firstChar && vowelArray.indexOf(firstChar) > -1) {
-                            outputPrepend("an");
+                            outputPrepend('an');
                         }
                         else if (firstChar) {
-                            outputPrepend("a");
+                            outputPrepend('a');
                         }
                         else {
                             outputPrepend(aOrAn);
                         }
                     }
-                    tokens[token] = output_1;
+                    tokens[token] = output;
                 }
-            };
-            for (var token in tokens) {
-                _loop_1(token);
             }
-            return tokens.join(" ").trim();
+            return tokens.join(' ').trim();
         };
-        this.update = function (update) {
-            var words = update.words, weight = update.weight;
+        this.update = (update) => {
+            const { words, weight } = update;
             if (words !== undefined) {
-                _this.words = words;
+                this.words = words;
             }
             if (weight !== undefined) {
-                _this.weight = weight;
+                this.weight = weight;
             }
         };
         this.words = words;
         this.weight = weight;
     }
-    return Word;
-}());
-exports.default = Word;
+}
+export default Word;
+//# sourceMappingURL=Word.js.map
