@@ -1,9 +1,4 @@
-"use strict";
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-var Word_1 = __importDefault(require("./Word"));
-var subTokens_1 = require("./subTokens");
-exports.SUBTOKENS = subTokens_1.SUBTOKENS;
-exports.default = Word_1.default;
+import Word from './Word';
+export { SUBTOKENS } from './subTokens';
+export default Word;
+//# sourceMappingURL=index.js.map
