@@ -1,9 +1,7 @@
 module.exports = {
-  "roots": [
-    "<rootDir>/src"
-  ],
-  "testEnvironment": "node",
-  "transform": {
-    "^.+\\.tsx?$": "ts-jest"
-  }
-}
+  roots: ['<rootDir>/src'],
+  testEnvironment: 'node',
+  transform: {
+    '^.+\\.tsx?$': 'ts-jest',
+  },
+};
