@@ -1,2 +1,2 @@
-import Bucket from "./Bucket";
+import Bucket from './Bucket';
 export default Bucket;
