@@ -1,7 +1,3 @@
-"use strict";
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
-Object.defineProperty(exports, "__esModule", { value: true });
-var Bucket_1 = __importDefault(require("./Bucket"));
-exports.default = Bucket_1.default;
+import Bucket from './Bucket';
+export default Bucket;
+//# sourceMappingURL=index.js.map
