@@ -1,8 +1,7 @@
-import Bucket from "../bucket";
-export declare class DuplicateNameError {
+import Bucket from '../bucket';
+export declare class DuplicateNameError extends Error {
     name: string;
-    message?: string;
+    message: string;
     duplicate: Bucket;
-    stack: any;
     constructor(message: string, bucket: Bucket);
 }
