@@ -1,4 +1,4 @@
-export { DeserialiseBucketError } from "./DeserialiseBucketError";
-export { DuplicateNameError } from "./DuplicateNameError";
-export { MissingBucketError } from "./MissingBucketError";
-export { ReservedWordError } from "./ReservedWordError";
+export { DeserialiseBucketError } from './DeserialiseBucketError';
+export { DuplicateNameError } from './DuplicateNameError';
+export { MissingBucketError } from './MissingBucketError';
+export { ReservedWordError } from './ReservedWordError';
