@@ -1,10 +1,5 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-var DeserialiseBucketError_1 = require("./DeserialiseBucketError");
-exports.DeserialiseBucketError = DeserialiseBucketError_1.DeserialiseBucketError;
-var DuplicateNameError_1 = require("./DuplicateNameError");
-exports.DuplicateNameError = DuplicateNameError_1.DuplicateNameError;
-var MissingBucketError_1 = require("./MissingBucketError");
-exports.MissingBucketError = MissingBucketError_1.MissingBucketError;
-var ReservedWordError_1 = require("./ReservedWordError");
-exports.ReservedWordError = ReservedWordError_1.ReservedWordError;
+export { DeserialiseBucketError } from './DeserialiseBucketError';
+export { DuplicateNameError } from './DuplicateNameError';
+export { MissingBucketError } from './MissingBucketError';
+export { ReservedWordError } from './ReservedWordError';
+//# sourceMappingURL=index.js.map
