@@ -1,14 +1,14 @@
 declare const _default: {
     BAD_COMMANDS: RegExp;
     BRACE_PAIRS: {
-        "{": string[];
-        "}": string[];
-        "[": string[];
-        "]": string[];
-        "(": string[];
-        ")": string[];
-        "<": string[];
-        ">": string[];
+        '{': string[];
+        '}': string[];
+        '[': string[];
+        ']': string[];
+        '(': string[];
+        ')': string[];
+        '<': string[];
+        '>': string[];
     };
     RESERVED_WORDS: string[];
 };
