@@ -1,4 +1,4 @@
-import Manager from "./Manager";
-export { default as VARS } from "./vars";
-export { default as CONST } from "./constants";
+import Manager from './Manager';
+export { default as VARS } from './vars';
+export { default as CONST } from './constants';
 export default Manager;
