@@ -1,4 +1,4 @@
-export { wordSummer } from "./wordSummer";
-export { getParentFromPath, splitPath, pathEnding, pathToTuple, tupleToPath } from "./namespacing";
-export { splitString } from "./splitter";
-export { checkFullToken, checkSubToken } from "./tokeniser";
+export { wordSummer } from './wordSummer';
+export { getParentFromPath, splitPath, pathEnding, pathToTuple, tupleToPath, } from './namespacing';
+export { splitString } from './splitter';
+export { checkFullToken, checkSubToken } from './tokeniser';
