@@ -1,15 +1,5 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-var wordSummer_1 = require("./wordSummer");
-exports.wordSummer = wordSummer_1.wordSummer;
-var namespacing_1 = require("./namespacing");
-exports.getParentFromPath = namespacing_1.getParentFromPath;
-exports.splitPath = namespacing_1.splitPath;
-exports.pathEnding = namespacing_1.pathEnding;
-exports.pathToTuple = namespacing_1.pathToTuple;
-exports.tupleToPath = namespacing_1.tupleToPath;
-var splitter_1 = require("./splitter");
-exports.splitString = splitter_1.splitString;
-var tokeniser_1 = require("./tokeniser");
-exports.checkFullToken = tokeniser_1.checkFullToken;
-exports.checkSubToken = tokeniser_1.checkSubToken;
+export { wordSummer } from './wordSummer';
+export { getParentFromPath, splitPath, pathEnding, pathToTuple, tupleToPath, } from './namespacing';
+export { splitString } from './splitter';
+export { checkFullToken, checkSubToken } from './tokeniser';
+//# sourceMappingURL=index.js.map
