@@ -1,20 +1,9 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.splitPath = function (path) {
-    return path.split(".");
-};
-exports.pathEnding = function (path) {
-    return exports.splitPath(path)[exports.splitPath(path).length - 1];
-};
-exports.getParentFromPath = function (path) {
-    return exports.splitPath(path).slice(0, -1).join(".");
-};
-exports.pathToTuple = function (path) {
-    return {
-        parent: exports.splitPath(path)[0],
-        child: path.split(exports.splitPath(path)[0] + ".")[1]
-    };
-};
-exports.tupleToPath = function (tuple) {
-    return tuple.join(".");
-};
+export const splitPath = (path) => path.split('.');
+export const pathEnding = (path) => splitPath(path)[splitPath(path).length - 1];
+export const getParentFromPath = (path) => splitPath(path).slice(0, -1).join('.');
+export const pathToTuple = (path) => ({
+    parent: splitPath(path)[0],
+    child: path.split(`${splitPath(path)[0]}.`)[1],
+});
+export const tupleToPath = (tuple) => tuple.join('.');
+//# sourceMappingURL=namespacing.js.map
