@@ -1,5 +1,5 @@
-export declare const splitString: (input: string, output?: string[]) => string[];
 export declare const findCommand: (input: string) => {
     commandChar: number;
     commandEnd: number;
 };
+export declare const splitString: (input: string, output?: string[]) => string[];
