@@ -1,2 +1,2 @@
-import Word from "../word";
+import Word from '../word';
 export declare const wordSummer: (words: Word[]) => number;
