@@ -1,5 +1,2 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.wordSummer = function (words) {
-    return words.reduce(function (sum, word) { return sum + word.weight; }, 0);
-};
+export const wordSummer = (words) => words.reduce((sum, word) => sum + word.weight, 0);
+//# sourceMappingURL=wordSummer.js.map
