@@ -1,3 +1,3 @@
-import Word from "./Word";
-export { SUBTOKENS } from "./subTokens";
+import Word from './Word';
+export { SUBTOKENS } from './subTokens';
 export default Word;
