@@ -1,17 +1,14 @@
 // This works where other, more comprehensible custom errors don't.
 
-export class MissingBucketError {
+export class MissingBucketError extends Error {
   public name: string;
-  public message?: string;
+  public message: string;
   public title: string;
-  public stack: any;
 
   constructor(message: string, title: string) {
-    this.name = "MissingBucketError";
-    this.message = message || "Bucket not found!";
+    super(message);
+    this.name = 'MissingBucketError';
+    this.message = message || 'Bucket not found!';
     this.title = title;
-    this.stack = (new Error()).stack;
   }
 }
-MissingBucketError.prototype = Object.create(Error.prototype);
-MissingBucketError.prototype.constructor = MissingBucketError;
