@@ -1,6 +1,7 @@
-import Manager from "./manager";
-export { default as Bucket } from "./bucket";
-export { DuplicateNameError, MissingBucketError } from "./errors";
-export { default as Word } from "./word";
+import Manager from './manager';
+
+export { default as Bucket } from './bucket';
+export { DuplicateNameError, MissingBucketError } from './errors';
+export { default as Word } from './word';
 
 export default Manager;
