@@ -1,22 +1,16 @@
-export const splitPath = (path: string): string[] => {
-  return path.split(".");
-}
+export const splitPath = (path: string): string[] => path.split('.');
 
-export const pathEnding = (path: string): string => {
-  return splitPath(path)[splitPath(path).length - 1];
-}
+export const pathEnding = (path: string): string =>
+  splitPath(path)[splitPath(path).length - 1];
 
-export const getParentFromPath = (path: string): string => {
-  return splitPath(path).slice(0,-1).join(".");
-}
+export const getParentFromPath = (path: string): string =>
+  splitPath(path).slice(0, -1).join('.');
 
-export const pathToTuple = (path: string): {parent: string, child: string} => {
-  return {
-    parent: splitPath(path)[0],
-    child: path.split(splitPath(path)[0] + ".")[1]
-  };
-}
+export const pathToTuple = (
+  path: string
+): { parent: string; child: string } => ({
+  parent: splitPath(path)[0],
+  child: path.split(`${splitPath(path)[0]}.`)[1],
+});
 
-export const tupleToPath = (tuple: string[]): string => {
-  return tuple.join(".");
-}
+export const tupleToPath = (tuple: string[]): string => tuple.join('.');
