@@ -1,4 +1,5 @@
-import Word from "./Word";
-export { SUBTOKENS } from "./subTokens";
+import Word from './Word';
+
+export { SUBTOKENS } from './subTokens';
 
 export default Word;
